Move avatar upload into settings page
